+++ message-bus-lib/src/main/java/wales/nhs/dhcw/msgbus/ProcessingResult.java
@@ -4,13 +4,14 @@ import java.util.Optional;
 
 public record ProcessingResult(
     boolean success,
-    Optional<String> errorReason
+    Optional<String> errorReason,
+    Optional<Boolean> retry
 ){
     public static ProcessingResult successful() {
-        return new ProcessingResult(true, Optional.empty());
+        return new ProcessingResult(true, Optional.empty(), Optional.empty());
     }
 
-    public static ProcessingResult failed(String errorReason) {
-        return new ProcessingResult(false, Optional.ofNullable(errorReason));
+    public static ProcessingResult failed(String errorReason, boolean retry) {
+        return new ProcessingResult(false, Optional.ofNullable(errorReason), Optional.ofNullable(retry));
     }
 }
